share button into account menu
